refactor: optimize delegation layers, deduplicate builders, and decouple server utils
- Inline pure delegation methods in HandoffService to use HandoffStorage directly.
- Extract duplicated execution request builders from plan.py and run.py into role_request_factory.py.
- Consolidate label/assignee normalization from governance.py into label_utils.py.
- Decouple webhook and port availability logic from server/app.py into webhook_utils.py and server_utils.py.
- Remove redundant auto_fix_branch from CheckService.
- Introduce singleton caching for IssueFlowService in issue_failure_service.py to prevent redundant database initializations.
- Update tests to reflect new _get_issue_flow_service patch boundaries.
