+++ src/vibe3/server/app.py
@@ -26,38 +26,12 @@ from vibe3.server.registry import (
     _start_async_serve,
     _validate_pid_file,
 )
+from vibe3.server.server_utils import ensure_port_available
 
 app = typer.Typer(
     help="Orchestra server: GitHub webhook receiver + heartbeat polling",
     no_args_is_help=True,
 )
-
-
-def _ensure_port_available(port: int) -> None:
-    """Raise typer.Exit if port is already in use."""
-    import socket
-
-    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
-        try:
-            # Use SO_REUSEADDR to be consistent with common server behavior
-            s.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-            s.bind(("0.0.0.0", port))
-        except OSError as e:
-            if e.errno in (48, 98):  # MacOS: 48, Linux: 98
-                typer.echo(
-                    f"\n[bold red]Error:[/] Port {port} is already in use.",
-                    err=True,
-                )
-                typer.echo(
-                    "Check if another Orchestra service is running on this port.",
-                    err=True,
-                )
-                typer.echo(
-                    "Use [bold]vibe3 serve stop[/] or specify [bold]--port[/].\n",
-                    err=True,
-                )
-                raise typer.Exit(1)
-            raise
 
 
 # --- Server Run Logic ---
@@ -223,7 +197,7 @@ def start(
         config.pid_file.unlink(missing_ok=True)
 
     # Pre-flight: Check if port is available
-    _ensure_port_available(config.port)
+    ensure_port_available(config.port)
 
     # Phase 1: FailedGate Preflight
     from vibe3.orchestra.failed_gate import FailedGate
