+++ src/vibe3/utils/label_utils.py
@@ -14,3 +14,16 @@ def normalize_labels(raw_labels: object) -> list[str]:
             if isinstance(name, str):
                 result.append(name)
     return result
+
+
+def normalize_assignees(raw_assignees: object) -> list[str]:
+    """Extract assignee logins from GitHub issue payload assignees field."""
+    if not isinstance(raw_assignees, list):
+        return []
+    assignees: list[str] = []
+    for item in raw_assignees:
+        if isinstance(item, dict):
+            login = item.get("login")
+            if isinstance(login, str) and login:
+                assignees.append(login)
+    return assignees
